// apps/api/src/generated/kysely/valuations/CurrencyIsoCode.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Represents the enum valuations.CurrencyIsoCode */
enum CurrencyIsoCode {
  CHF = 'CHF',
  EUR = 'EUR',
  GBP = 'GBP',
  NOK = 'NOK',
  SEK = 'SEK',
  USD = 'USD',
  DKK = 'DKK',
};

export default CurrencyIsoCode;